Change behavior of update()
update() now passes `valIfNotSet` argument to the updater function if provided.
Previously, `valIfNotSet` was used to update the key path when it did not exist.

Now the updater function will always be called. Previously, it was not called
when the key path did not exist, which caused updates to fail silently.

setDefault() should now be used for initialization of values instead of
update().